requirements.txt
aiofiles==24.1.0
boto3==1.36.2
botocore==1.36.2
fastapi==0.115.7
langchain_core==0.3.32
langchain_groq==0.2.3
langchain_ollama==0.2.2
langchain_openai==0.3.2
langgraph==0.2.68
pandas==2.2.3
pdf2image==1.17.0
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.41.1
tenacity==8.5.0
typing_extensions==4.12.2
uvicorn==0.34.0
python-multipart
openpyxl